RISC-V: Make jump benefits signed
A bug exists whereby instructions for jump tables are being inserted by
the linker, but the table not being instantiated has occurred due to a
comparison between the jump table benefit (as an unsigned value), and
the end of table analysis marker (signed -1) being treated as an
unsigned comparison and incorrectly noting that there is no benefit,
causing the table to be omitted. This patch changes the values holding
size differences to be signed, allowing these comparisons to behave
correctly, and correctly generate the table.

Signed-off-by: Simon Cook <simon.cook@embecosm.com>